+++ src/k8s-extension/HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ===============
 
+1.0.0
+++++++++++++++++++
+* Switch to GA api-version of Extensions (2021-09-01)
+* Support Extensions PATCH
+* Enable Dapr extension type
+* Enable ManagedClusters clusterType
+
+0.7.1
+++++++++++++++++++
+* Fix DF resource manager endpoint check
+
 0.7.0
 ++++++++++++++++++
 * Enable identity by default for extensions